moved a few component that should have been in the views folder
