+++ Web/utkeyrslukerfi-web/src/App.js
@@ -2,20 +2,20 @@ import './styles/navbar.css'
 import './styles/main.css'
 import { Route, Switch } from 'react-router-dom'
 import { useDispatch, useSelector } from 'react-redux'
-import Home from './views/Home'
-import Users from './views/Users'
-import User from './views/User'
-import Deliveries from './views/Deliveries'
-import Container from './components/Container'
 import React, { useEffect } from 'react'
 import UpdatePasswordForm from './components/UpdatePasswordForm'
+import Container from './components/Container'
 import Login from './components/Login'
-import Delivery from './components/Delivery'
-import Package from './components/Package'
 import Navbar from './components/Navbar'
+import { getLogin } from './actions/loginActions'
+import User from './views/User'
+import Home from './views/Home'
+import UsersList from './views/UsersList'
+import DeliveriesList from './views/DeliveriesList'
+import UpdateDelivery from './views/UpdateDelivery'
+import Package from './views/Package'
 import CreateUser from './views/CreateUser'
 import NotFound from './views/NotFound'
-import { getLogin } from './actions/loginActions'
 
 const App = () => {
   const token = useSelector(({ login }) => login.token)
@@ -38,11 +38,11 @@ const App = () => {
       <Container>
         <Switch>
           <Route exact path='/' component={Home} />
-          <Route exact path='/users' component={Users} />
+          <Route exact path='/users' component={UsersList} />
           <Route exact path='/users/create' component={CreateUser} />
           <Route exact path='/users/:id' component={User} />
-          <Route exact path='/deliveries' component={Deliveries} />
-          <Route exact path='/deliveries/:id' component={Delivery} />
+          <Route exact path='/deliveries' component={DeliveriesList} />
+          <Route exact path='/deliveries/:id' component={UpdateDelivery} />
           <Route exact path='/deliveries/:delid/packages/:id' component={Package} />
           <Route exact path='*' component={NotFound} />
         </Switch>
